ath10k: Don't initialize devices asynchronously
OpenWrt requires all PHYs to be initialized to create the configuration files
during bootup. ath10k violates this because it delays the creation of the PHY
to a not well defined point in the future.

Forcing the work to be done immediately works around this problem but may also
delay the boot when firmware images cannot be found.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>