ath10k: Support custom tx and per-frame rate-ctrl for 10.4 CT firmware.
10.4 firmware in general has a feature (that may be compiled out in stock FW)
that attempts to allow one to specify a per-frame tx rate.  It uses an extra
32-bit word in the tx descriptor, which ath10k does not support.  But, we can play
some tricks and use the peer-id field since ath10k always sets it to 0xFFFF.  In
off-chan tx, the high 16 bits are the freq, but we can still use bit 0x20000000
since the freq would never be high enough to need to set that bit.  This bit is
used by the firmware to flag whether user has specified a particular rate or not.

In addition, while coding this up, I realized my similar hack for 10.1 firmware
would likely break the off-chan tx due to not properly realizing the freq could
be set in the high bits.  So, I changed that API as well.  A new CT 10.1 firmware
will be needed to work with custom tx rates once this patch is applied.

Signed-off-by: Ben Greear <greearb@candelatech.com>