wireless: Hack to fix AP + STA on same machine on DFS channels.
HACK:  Work around problem where you have AP on DFS channel and then
STA on different radio connects on same channel.  That causes regdom to change
(or the code isn't smart enough to realize it didn't really change),
because STA gets regdom from its AP, causing CAC to restart,
which kills the AP interface before CAC can ever be finished.
This is the one path that hits in my system, there are other places that may
need latching too, and/or there is probably a better way to fix this.

Signed-off-by: Ben Greear <greearb@candelatech.com>